backend/requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
billiard==3.5.0.5
celery==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.5.0
django-crontab==0.7.1
django-dotenv==1.4.2
django-timezone-field==4.0
djangorestframework==3.12.1
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
kombu==4.6.11
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
zipp==3.3.1
